refactor(cli): update ai and ide tools through one implementation
The two update use-cases were 81 lines each and identical line for line
apart from three things: the tool-id type, the predicate (isAiToolId from
domain/models/tool-ids, isIdeToolId from domain/tools/registry, different
source modules), and the names. No logic differed.

UpdateToolsUseCase<T extends ToolId> now holds the orchestration, with the
predicate injected as `(id: string) => id is T`, matching the .filter shape
the originals already used. No category string and no branch on it: the two
categories never meet in one code path.

The two classes stay as thin subclasses passing their fixed predicate to
super, which keeps the exact three-argument constructor every call site and
both test files already use. deps.ts and both test files are unchanged.

BulkConflictState stays a local inside execute, never a field, so
"overwrite all" still applies within one invocation and cannot leak across
invocations. Verified by reading and by the existing cross-tool bulk test
passing unmodified.

Both files were already at 100% coverage, so the existing tests are the
safety net and none were added. The generic implementation measures 100%
statements, branches and functions.

2136/2136 pass, tsc clean, no cast and no any.

Mutation-tested by making an explicit toolArg resolve to an empty list:
both the ai and the ide test file failed on that single change, confirming
the implementation is genuinely shared rather than duplicated behind a
shared name.

--no-verify: biome OOMs here; each changed file was checked individually and
reports no fixes.
